Created Information Box
created new src information.lua, made information object, constructor, and draw function. Information Box draws to right side of screen in front of the galaxy map. Also changed spelling error in a comment Galaxy Type in the galaxy source. Added a setColor function into love.draw in the main file because the gameboard:draw does not have a set color, meaning when the game repeats the color becomes the same as the info box.